% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{BDSS}
\alias{BDSS}
\alias{BDSS-package}
\title{Package containing internal BaDaSS convenience functions}
\description{
This package contains the following functions
\enumerate{
  \item SummaryTable
  \item require.anyway
  \item GetCommonPackages
  }
}
\author{
Dominic Laroche,
Chang Xu,
Brian Mannakee

Maintainer: Dominic Laroche \email{dominic.laroche@contractors.roche.com}
Depends: xtable, rtf, ggplot2
}

