% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{BDSSRecurrenceExample}
\alias{BDSSRecurrenceExample}
\title{A Sample Recurrence Dataset}
\format{A data frame with 200 rows and 6 variables}
\description{
This is a randomly generated dataset containing
recurrence data on 200 patients on one of two treatments
based on their age. Mostly used to test the function SummaryTable.
The variables are as follows:
}
\details{
\itemize{
  \item Age. age of patients, mean 60
  \item Recur. did the patient recur (recurred,not recurred)
  \item treated. did the patient recieve treatment 1 (Treated, Not Treated)
  \item trt2. did the patient recieve treatment 2 (Treatment2, No Trt2)
  \item meas. Age - 30.
  \item single. all patients get single treatment. (Yes)
}
}
\keyword{datasets}

