% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggSurvival}
\alias{ggSurvival}
\title{Kaplan Meier plots with the ggplot package}
\usage{
ggSurvival(s, CI = FALSE, spot.cens = FALSE, surv.col = 'gg.def',
cens.col = 'red', lty.est = 1, lty.ci = 2, lwd.est=1, lwd.ci=1,
cens.shape = 3, black.white = TRUE, xlab = 'Time', ylab = 'Survival',
 main = '',stratName=NULL, groupNames=NULL, groupNameLevels=NULL)
}
\arguments{
\item{s}{A object of class 'survfit'}

\item{CI}{Logical indicating whether to plot confidence intervals as defined in the survfit object}

\item{spot.cens}{Logical indicating whether to add tick marks to censor points}

\item{surv.col}{A vector of colors for the survival curves}

\item{cens.col}{A color for the censor ticks}

\item{lty.est}{A vector of line types for the survival curves}

\item{lty.ci}{A vector of line types for the confidence intervals}

\item{lwd.est}{A vector of line widths for the survival curves}

\item{lwd.ci}{A vector of line widths for the confidence intervals}

\item{cens.chape}{A character string or number indicating what shape the censor indicator should be.}

\item{black.white}{Logical indicating whether to use \code{theme_bw}}

\item{xlab,ylab,main}{Same as ggplot2 and base}

\item{stratName}{A character string which overrides the name of the stratification variable (optional for prettyification).}

\item{groupNames}{A character vector renaming the levels of the stratification variable.  These must be in the same order as the original factor levels (optional).}

\item{groupNameLevels}{A character vector of the stratification variable levels in the order which you want them to appear.}

\item{ylimit}{a vector giving the limits for plotting the y axis (default to c(0,1)).}
}
\description{
ggSurvplot creates a ggplot object from a '\code{survfit}' object.  The resulting plot can be extended using any of ggplot's many tools
}
\details{
The plot is constructed from the survfit object by accessing the internal structure of the fit and plotting a curve for each strata and confidence interval if desired.
}
\examples{
\donttest{
require(survival)
require(ggplot2)
data(BDSSRecurrenceExample)
dat <- BDSSRecurrenceExample
dat$Rec<-ifelse(dat$Recur=="recurred",1,0)
sv<-Surv(dat$time,dat$Rec)
sv.fit<-survfit(sv~treated,data=dat)
ggSurvival(sv.fit,stratName="Treatment")+theme(legend.position="bottom")
}
}
\author{
Dominic LaRoche (adapted from Edwin Thoen)
}

