% Generated by roxygen2 (4.0.1): do not edit by hand
\name{require.anyway}
\alias{require.anyway}
\title{require.anyway}
\usage{
require.anyway(pkg,verbose=TRUE)
}
\arguments{
\item{pkg}{the package to be loaded}

\item{verbose}{logical, whether to print confirmation that package loaded correctly}
}
\description{
checks the local system to see if a package
is installed, and if not, installs it
}
\details{
the function wraps a call to install.packages
that uses system defaults. \code{install.packages(pkg)}
}
\author{
Chang Xu
}

