% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SummaryTable}
\alias{SummaryTable}
\title{Summary table from recurrence data in a data frame}
\usage{
SummaryTable(data,rowvars,colvar,row.names=NULL,cont.vars=NULL,percent=TRUE,output=c("latex","matrix"), reorder.columns=NULL,
cap=NULL,shortcap=NULL,lab=NULL,rnd.val=1,var.totals=FALSE,missing=TRUE,outfile="~/table1output.rtf",rtf.font.size=11,
col.widths=NULL, col.justify=('C'), header.col.justify=('C'),
rtf.row.names=FALSE, NA.string="-", space.before=NULL, space.after=NULL,
rmark.split=Inf)
}
\arguments{
\item{data}{matrix or data frame with the row and column variables}

\item{rowvars}{a character vector of quoted variable names or a numeric vector
of categorical variables}

\item{colvar}{a quoted variable name or a number identifying the column variable
in the data}

\item{row.names}{an optional character vector of row names if you want to change
the displayed name for row variables. \code{len(row.names)} must equal \code{len(rowvars)}}

\item{cont.vars}{a character vector of variable names, or a number vector of column indices}

\item{percent}{logical indicating whether to print percentages for each row}

\item{output}{a vector of outputs including one or more of \code{latex,matrix,rtf}}

\item{reorder.columns}{optional character string specifying the column level
to appear first in the table}

\item{cap}{character string with the desired table caption}

\item{shortcap}{Character string with desired short caption to use in list of tables.}

\item{lab}{character string with table label}

\item{rnd.val}{number of digits to round to}

\item{var.totals}{logical indicating whether to print totals for each row variable}

\item{missing}{logical indicating whether to create a row within each
row variable for NA values}

\item{outfile}{if \code{output='rtf'}, a file path to save the output}

\item{rtf.font.size}{if \code{output='rtf'}, the size of the font in the rtf table}

\item{col.widths}{if \code{output='rtf'}, a vector of column widths}

\item{col.justify}{if \code{output='rtf'}, a vector of column alignments}

\item{header.col.justify}{if \code{output='rtf'}, a vector of alingments for column headers}

\item{rtf.row.names}{if \code{output='rtf'}, logicial indicating whether to print row names}

\item{NA.string}{if \code{output='rtf'}, string used to replace NA}

\item{space.before}{if \code{output='rtf'}, see package rtf for details}

\item{space.after}{if \code{output='rtf'}, see package rtf for details}
}
\description{
SummaryTable automatically creates summary tables in latex format.
The function calculates the frequency and percent(optional) of
each level of each categorical row variable(rowvars) for each level
of each column variable(colvars).
}
\details{
The core function of the table builder consists of repeated calls
to a \code{tapply} statment:
\code{table <- tapply(data$count, list(rowlist[[i]],data[ ,colvar]),sum)}

where data$count is a function defined variable with a \code{1} in each
row for tabulating frequencies. \code{rowlist[[i]]} represents each row variable
and \code{data[ ,colvar]} is the column variable.
Currently the function does not support \code{NA} values in the column variable.
}
\examples{
\donttest{
data(BDSSRecurrenceExample)
fr <- BDSSRecurrenceExample
SummaryTable(data=fr,rowvars=c("treated","trt2","single"),colvar="Recur",cont.vars=c("Age","meas"))
}
}
\author{
Dominic LaRoche
}

